Perf-Trending Categories to use partial rendering and suspense skeletons
